requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
ezdxf>=1.1.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
openpyxl>=3.1.0
jinja2>=3.1.0
python-multipart>=0.0.9
